IMU and attitude: general clean-up
