+++ mavros/src/plugins/hil.cpp
@@ -127,13 +127,6 @@ private:
 	void state_quat_cb(const mavros_msgs::HilStateQuaternion::ConstPtr &req) {
 		mavlink::common::msg::HIL_STATE_QUATERNION state_quat;
 
-		// get current ENU FCU orientation
-		tf::quaternionMsgToEigen(m_uas->get_attitude_orientation(), enu_orientation);
-
-		// body frame rotations must be aware of current attitude of the vehicle
-		auto q_ned_current = ftf::transform_orientation_enu_ned(
-					ftf::transform_orientation_baselink_aircraft(enu_orientation));
-
 		state_quat.time_usec = req->header.stamp.toNSec() / 1000;
 		auto q = ftf::transform_orientation_baselink_aircraft(
 					ftf::transform_orientation_enu_ned(
@@ -151,14 +144,13 @@ private:
 		state_quat.alt = req->geo.altitude * 1E3;
 		state_quat.ind_airspeed = req->ind_airspeed * 1E2;
 		state_quat.true_airspeed = req->true_airspeed * 1E2;
-		// angular velocity - WRT body frame
+		// WRT world frame
 		auto ang_vel = ftf::transform_frame_enu_ned(
 					ftf::transform_frame_baselink_aircraft(
-						ftf::transform_frame_aircraft_ned(Eigen::Vector3d(
-								req->angular_velocity.x,
-								req->angular_velocity.y,
-								req->angular_velocity.z), q_ned_current)));
-		// linear velocity - WRT world frame
+						Eigen::Vector3d(
+							req->angular_velocity.x,
+							req->angular_velocity.y,
+							req->angular_velocity.z)));
 		auto lin_vel = ftf::transform_frame_enu_ned<Eigen::Vector3d>(
 					Eigen::Vector3d(
 						req->linear_velocity.x,
@@ -232,27 +224,18 @@ private:
 	void sensor_cb(const mavros_msgs::HilSensor::ConstPtr &req) {
 		mavlink::common::msg::HIL_SENSOR sensor;
 
-		// get current ENU FCU orientation
-		tf::quaternionMsgToEigen(m_uas->get_attitude_orientation(), enu_orientation);
-
-		// body frame rotations must be aware of current attitude of the vehicle
-		auto q_ned_current = ftf::transform_orientation_enu_ned(
-					ftf::transform_orientation_baselink_aircraft(enu_orientation));
-
 		sensor.time_usec = req->header.stamp.toNSec() / 1000;
-		// acceleration values - WRT world frame
+		// WRT world frame
 		auto acc = ftf::transform_frame_baselink_aircraft(
 					Eigen::Vector3d(
 						req->acc.x,
 						req->acc.y,
 						req->acc.z));
-		// angular values - WRT body frame
 		auto gyro = ftf::transform_frame_baselink_aircraft(
-					ftf::transform_frame_aircraft_ned(Eigen::Vector3d(
-							req->gyro.x,
-							req->gyro.y,
-							req->gyro.z), q_ned_current));
-		// heading values - WRT world frame
+					Eigen::Vector3d(
+						req->gyro.x,
+						req->gyro.y,
+						req->gyro.z));
 		auto mag = ftf::transform_frame_baselink_aircraft<Eigen::Vector3d>(
 					Eigen::Vector3d(
 						req->mag.x,
